Remove lwjgl dependencies from GUIManager by making it into an interface
Add headless GUIManager
